relocate beacon blocks by range
