+++ packages/beacon-node/src/network/network.ts
@@ -12,10 +12,9 @@ import {altair, eip4844, Epoch, phase0} from "@lodestar/types";
 import {IMetrics} from "../metrics/index.js";
 import {ChainEvent, IBeaconChain, IBeaconClock} from "../chain/index.js";
 import {BlockInput, BlockInputType, getBlockInput} from "../chain/blocks/types.js";
-import {ckzg} from "../util/kzg.js";
 import {INetworkOptions} from "./options.js";
 import {INetwork} from "./interface.js";
-import {ReqRespBeaconNode, ReqRespHandlers} from "./reqresp/index.js";
+import {ReqRespBeaconNode, ReqRespHandlers, doBeaconBlocksMaybeBlobsByRange} from "./reqresp/index.js";
 import {Eth2Gossipsub, getGossipHandlers, GossipHandlers, GossipTopicTypeMap, GossipType} from "./gossip/index.js";
 import {MetadataController} from "./metadata.js";
 import {FORK_EPOCH_LOOKAHEAD, getActiveForks} from "./forks.js";
@@ -228,77 +227,7 @@ export class Network implements INetwork {
     peerId: PeerId,
     request: phase0.BeaconBlocksByRangeRequest
   ): Promise<BlockInput[]> {
-    // TODO EIP-4844: Assumes all blocks in the same epoch
-    // TODO EIP-4844: Ensure all blocks are in the same epoch
-    if (this.config.getForkSeq(request.startSlot) < ForkSeq.eip4844) {
-      const blocks = await this.reqResp.beaconBlocksByRange(peerId, request);
-      return blocks.map((block) => getBlockInput.preEIP4844(this.config, block));
-    }
-
-    // Only request blobs if they are recent enough
-    else if (
-      computeEpochAtSlot(request.startSlot) >=
-      this.chain.clock.currentEpoch - this.config.MIN_EPOCHS_FOR_BLOBS_SIDECARS_REQUESTS
-    ) {
-      // TODO EIP-4844: Do two requests at once for blocks and blobs
-      const blocks = await this.reqResp.beaconBlocksByRange(peerId, request);
-      const blobsSidecars = await this.reqResp.blobsSidecarsByRange(peerId, request);
-
-      const blockInputs: BlockInput[] = [];
-      let blobSideCarIndex = 0;
-      let lastMatchedSlot = -1;
-
-      // Match blobSideCar with the block as some blocks would have no blobs and hence
-      // would be omitted from the response. If there are any inconsitencies in the
-      // response, the validations during import will reject the block and hence this
-      // entire segment.
-      //
-      // Assuming that the blocks and blobs will come in same sorted order
-      for (let i = 0; i < blocks.length; i++) {
-        const block = blocks[i];
-        let blobsSidecar: eip4844.BlobsSidecar;
-
-        if (blobsSidecars[blobSideCarIndex]?.beaconBlockSlot === block.message.slot) {
-          blobsSidecar = blobsSidecars[blobSideCarIndex];
-          lastMatchedSlot = block.message.slot;
-          blobSideCarIndex++;
-        } else {
-          // Quick inspect if the blobsSidecar was expected
-          const blobKzgCommitmentsLen = (block.message.body as eip4844.BeaconBlockBody).blobKzgCommitments.length;
-          if (blobKzgCommitmentsLen !== 0) {
-            throw Error(
-              `Missing blobsSidecar for blockSlot=${block.message.slot} with blobKzgCommitmentsLen=${blobKzgCommitmentsLen}`
-            );
-          }
-          blobsSidecar = {
-            beaconBlockRoot: this.config.getForkTypes(block.message.slot).BeaconBlock.hashTreeRoot(block.message),
-            beaconBlockSlot: block.message.slot,
-            blobs: [],
-            kzgAggregatedProof: ckzg.computeAggregateKzgProof([]),
-          };
-        }
-        blockInputs.push(getBlockInput.postEIP4844(this.config, block, blobsSidecar));
-      }
-
-      // If there are still unconsumed blobs this means that the response was inconsistent
-      // and matching was wrong and hence we should throw error
-      if (blobsSidecars[blobSideCarIndex] !== undefined) {
-        throw Error(
-          `Unmatched blobsSidecars, blocks=${blocks.length}, blobs=${
-            blobsSidecars.length
-          } lastMatchedSlot=${lastMatchedSlot}, pending blobsSidecars slots=${blobsSidecars
-            .slice(blobSideCarIndex)
-            .map((blb) => blb.beaconBlockSlot)}`
-        );
-      }
-      return blockInputs;
-    }
-
-    // Post EIP-4844 but old blobs
-    else {
-      const blocks = await this.reqResp.beaconBlocksByRange(peerId, request);
-      return blocks.map((block) => getBlockInput.postEIP4844OldBlobs(this.config, block));
-    }
+    return doBeaconBlocksMaybeBlobsByRange(this.config, this.reqResp, peerId, request, this.clock.currentEpoch);
   }
 
   async beaconBlocksMaybeBlobsByRoot(peerId: PeerId, request: phase0.BeaconBlocksByRootRequest): Promise<BlockInput[]> {
