+++ packages/beacon-node/src/network/reqresp/index.ts
@@ -1,2 +1,3 @@
 export * from "./ReqRespBeaconNode.js";
 export * from "./interface.js";
+export * from "./doBeaconBlocksMaybeBlobsByRange.js";
